Add support for CJS to fix #106
